+++ test/board/boardtest.cpp
@@ -50,8 +50,8 @@ Returns a (heap-allocated) vector of all valid, unique Board Squares.
 */
 std::unique_ptr<std::vector<Square>> getAllSquares() {
     auto squares = std::unique_ptr<std::vector<Square>>(new std::vector<Square>);
-    for (int irow = 0; irow < Board::WIDTH; ++irow) {
-        for (int icol = 0; icol < Board::WIDTH; ++icol) {
+    for (int irow = 0; irow < static_cast<int>(Board::WIDTH); ++irow) {
+        for (int icol = 0; icol < static_cast<int>(Board::WIDTH); ++icol) {
             squares->push_back(Square(irow, icol));
         }
     }
@@ -291,7 +291,7 @@ TEST(BoardTest, GetOccupiedSquaresTest) {
             }
             PieceColor color = static_cast<PieceColor>(icolor);
             size_t num_squares = board.getOccupiedSquares(color, occupied_buffer.start());
-            for (int isquare = 0; isquare < num_squares; ++isquare) {
+            for (int isquare = 0; isquare < static_cast<int>(num_squares); ++isquare) {
                 Square square = occupied_buffer.get(isquare);
                 // make sure this square still exists in the set
                 ASSERT_NE(square_set.end(), square_set.find(square));
@@ -301,7 +301,7 @@ TEST(BoardTest, GetOccupiedSquaresTest) {
                 square_set.erase(square);
             }
             // make sure all squares were accounted for
-            ASSERT_EQ(0, square_set.size());
+            ASSERT_EQ(static_cast<std::size_t>(0), square_set.size());
         }
     }
 }
